+++ src/Rule/HaveIBeenPwned.php
@@ -47,9 +47,6 @@ final class HaveIBeenPwned implements RuleInterface
         if ($max !== null && $max < $min) {
             throw new InvalidArgumentException('Max cannot be smaller than min.');
         }
-        if ($min === 0 && $max === null) {
-            throw new InvalidArgumentException('Min cannot be zero when max is unconstrained.');
-        }
 
         $this->min = $min;
         $this->max = $max;