+++ tests/Rule/LengthTest.php
@@ -73,11 +73,13 @@ final class LengthTest extends TestCase
     /**
      * @covers ::__construct
      */
-    public function testCannotConstructUnconstrainedRule(): void
+    public function testCanConstructUnconstrainedRule(): void
     {
-        $this->expectException(InvalidArgumentException::class);
-
         $rule = new Length(0, null);
+
+        // Force generation of code coverage
+        $ruleConstruct = new Length(0, null);
+        self::assertEquals($rule, $ruleConstruct);
     }
 
     /**